Make all the OIB commands return result.
Currently, as void functions, all OIB commands return nothing at the output, while in fact they should report the results.
Due to this problem, some commands did some dirty hacks to report the results in ad-hoc manner.
A clear example to this issue is how many h2fmi commands report the result with messages like "fmi: Command completed with result ..."

This patch fixes the problem by making all the OIB commands return an int value as result.
The result is then reported in the output universally, in an unified manner.
Thanks to this, all the ad-hoc reports on the results become redundant, therefore removed.